requirements.txt
Pillow>=10.0.0
Flask>=3.0.0
Flask-Cors>=4.0.0
python-dotenv>=1.0.1
requests>=2.32.0
